requirements.txt
# Core LLM & orchestration
groq>=0.9.0
langchain-core>=0.2.0
langchain>=0.2.0
langchain-text-splitters
langchain-community>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0

# Vector store
faiss-cpu>=1.7.4

# Embeddings (local, no API needed)
sentence-transformers>=3.0.0
torch>=2.0.0

# PDF text extraction
pypdf>=4.0.0
pdfplumber>=0.10.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0

# UI
streamlit>=1.32.0

# LangSmith tracing
langsmith>=0.1.0